Wrote Command Framework
this includes:
GyroDrive (AHRS-based PID Command), TeleopDrive Frame (Zak needs to write the command), solenoid commands, and a hopefully functioning robot.java, along with an autonomous drive a distance command (unfinished need to calculate ticks per distance).

Leaveline needs a tune, that's about it. dTerm is also acting up, so we need to figure out what's wrong. Added an end() which means I can run it in a commandgroup.